doc note about strings as regexen
Add a note in the String Operators section about strings being usable as
regexen, and a note in the split example in particular.

In #17, Schwern commented about it not being obvious to users that
qr// is needed where // would normally be used, but it looks like
previous people polishing docs clarified that already (thanks doc
polishers!).  So this note about strings as regexes closes #17.