s() doesn't work sensibly #12 / return modified string
Continuing with the "I didn't get around to ripping operators out" theme,
s() now continues to return true/false when used in an if(), !, ?:, etc
(boolean context as detected by Want), but in any and all other scalar
contexts, returns the modified string instead, which will often (but not
always!) pass as a true value.

This pulls in Want, which could potentially break things, but CPANTS
for it looks pretty durn good.

Closes #12.